utilites/recog_test2/requirements.txt
# recognise img
ultralytics==8.4.21
# mod image
pillow==11.3.0
# dataset read, filter
pandas==3.0.1
